refactor(oci)!: Move `PlatformChecksum` into utils package
